Entire input now has HTML tags sterilized on read-in
