+++ eddymc/scale/scale_html_writer.py
@@ -5,7 +5,6 @@
 """ This module creates the HTML output for the SCALE data """
 
 # Standard library imports
-import os
 import datetime
 try:
     import importlib.resources as pkg_resources
@@ -13,7 +12,7 @@ except ImportError:
     import importlib_resources as pkg_resources
 # Third party imports
 from jinja2 import Template
-#local imports
+# local imports
 try:
     from .. import static
 except ValueError:
@@ -30,23 +29,6 @@ def get_css():
     return inline_css
 
 
-def sanitize_input(scale_input):
-    """Replaces any < and > characters in the scale output with html codes &lt and &gt to stop
-    them from being interpreted as html tags.
-
-    Args: mcnp_input [list]: The mcnp input file
-    Returns: san_mcnp_input [list]: The sanitized version of the mcnp input
-    """
-    san_scale_input = []
-    for line in scale_input:
-        if "<" in line:
-            line = line.replace("<","&lt")
-        if ">" in line:
-            line = line.replace(">", "&gt")
-        san_scale_input.append(line)
-    return san_scale_input
-
-
 def get_html(case):
     """Get the html output from the jinja template
 
@@ -57,14 +39,13 @@ def get_html(case):
         html (str): the html output as a single string
     """
     inline_css = get_css()
-    scale_input = sanitize_input(case.scale_input)
     html_template = pkg_resources.read_text(static, 'SCALE_template.html')
     template = Template(html_template)
     # render template as a unicode string
     html = template.render(
         filename=case.filepath,
         case_name=case.name,
-        #logo=f"{os.getcwd()}/static/logo.png",
+        # logo=f"{os.getcwd()}/static/logo.png",
         inline_css=inline_css,
         rundate=case.rundate,
         runtime=case.runtime,
@@ -73,10 +54,6 @@ def get_html(case):
         scaling_factor=case.scaling_factor,
         tally_list=case.tally_list,
         mixture_list=case.mixture_list,
-        input_deck=scale_input,
+        input_deck=case.scale_input,
         )
     return html
-
-
-
-
