+++ tests/mcnp/test_mcnp_html_writer.py
@@ -3,7 +3,7 @@
 or add a configuration in pycharm
 """
 
-from eddymc.mcnp.mcnp_html_writer import get_css, sanitize_list
+from eddymc.mcnp.mcnp_html_writer import get_css
 
 
 def test_get_css():
@@ -13,20 +13,3 @@ def test_get_css():
     # assert
     assert type(actual_css) == str
     assert actual_css is not ''
-
-
-def test_sanitize_list():
-    # arrange
-    test_text = [
-        "This is ordinary text",
-        "This has a <h1> </h1> symbol in it",
-        "This has a \" symbol in it",
-        "This has a & in it.",
-        ]
-    # act
-    sanitized_text = sanitize_list(test_text)
-    # assert
-    assert sanitized_text[0] == "This is ordinary text"
-    assert sanitized_text[1] == "This has a &lt;h1&gt; &lt;/h1&gt; symbol in it"
-    assert sanitized_text[2] == "This has a &#34; symbol in it"
-    assert sanitized_text[3] == "This has a &amp; in it."
